+++ .eslintignore
@@ -10,3 +10,4 @@ report/
 template/
 test/test-cases/*/*
 !test/test-cases/*/*.test.js
+fixtures/*/*
