+++ .npmignore
@@ -1,1 +1,4 @@
 test/
+tests/
+fixtures/
+docs/
