+++ .npmrc
@@ -1,1 +1,2 @@
-package-lock=false
+# `treat` needs this to be hoisted
+public-hoist-pattern[]="virtual-resource-loader"
