+++ fixtures/braid-design-system/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@fixtures/braid-design-system",
+  "private": true,
+  "exports": {
+    "./sku.config.js": "./sku.config.js"
+  },
+  "dependencies": {
+    "braid-design-system": "^31.0.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "sku": "workspace:*"
+  }
+}
