+++ fixtures/sku-webpack-plugin/main.js
