--- test/test-cases/braid-design-system/app/README.md
@@ -1,2 +1,0 @@
-We've symlinked React and braid-design-system dependencies into ./node_modules because our package
-resolution helpers and webpack-node-externals depend on it.
