--- test/test-cases/braid-design-system/app/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "braid-design-system": "*"
-  }
-}
